SHA256 crashed due to fixed length buffers for SHA1
This fix isn't ideal, but is working for SHA256. The changes in create.c
addresses a crash by buffer overrun, changing local variables.
But there is another crash, when the operation completes. The change in
netpgpdigest.g fixes this.
A better way is to dynamically provide buffers with respect to the
chosen hash algorithm.